feat: introduce stricter lint configuration and clippy.toml
We spend a lot of time on lints and attributes for linting, and manually tuning our own lints for, frankly, low signal-to-noise stuff (`clippy::too_many_arguments`).

Examples:
- https://github.com/0xMiden/miden-vm/pull/2422
- https://github.com/0xMiden/miden-vm/pull/2317
- https://github.com/0xMiden/miden-vm/pull/1826

We also inflict nits about this to each other in PR review, e.g. :
- https://github.com/0xMiden/miden-vm/pull/2366#discussion_r2561945992

This PR is about sparing us cycles on this easy-to-mechanize stuff. It introduces bog-standard lint configuration through a .clippy.toml and .cargo/config.toml with a couple of extra-mild lints, and adjusts the code base in function.

<details><summary> Activated Clippy Lints </summary>

Clippy Lints

- **clippy::all**: Enables all default clippy lints that are not specifically enabled by other lints
- **clippy::assertions_on_result_states**: Warns against using assertions on Result types instead of proper error handling
- **clippy::cast_lossless**: Identifies cast conversions between numerical types that could be replaced by safer, more explicit conversion functions like Into or From.
- **clippy::checked_conversions**: Checks for explicit checked conversions that could be made more idiomatic
- **clippy::dbg_macro**: Warns about using the `dbg!` macro in production code
- **clippy::disallowed_methods**: Allows configuration of methods that should not be used
- **clippy::filter_map_next**: Checks for usage of `filter_map().next()` that could be `find()`
- **clippy::flat_map_option**: Checks for `flat_map(Option)` usage that could be more efficiently written
- **clippy::inefficient_to_string**: Warns about inefficient `to_string()` conversions
- **clippy::large_stack_arrays**: Warns about large arrays allocated on the stack that could cause stack overflow
- **clippy::large_types_passed_by_value**: Warns when large types are passed by value instead of by reference
- **clippy::macro_use_imports**: Discourages the use of `macro_use` in favor of more explicit imports
- **clippy::manual_assert**: Suggests using the `assert!` macro instead of manual assertion implementations
- **clippy::manual_ok_or**: Suggests using the `ok_or()` method instead of manual implementations
- **clippy::map_flatten**: Checks for usage of `map().flatten()` that could be replaced with `flat_map()`
- **clippy::map_unwrap_or**: Checks for usage of `map().unwrap_or()` that could be more safely written
- **clippy::match_wild_err_arm**: Warns about wildcard patterns in error handling match arms
- **clippy::needless_borrow**: Detects unnecessary borrows that can be removed
- **clippy::option_option**: Warns about nested `Option<Option<T>>` types that could be simplified
- **clippy::same_functions_in_if_condition**: Detects repeated function calls in if conditions that can be cached
- **clippy::single_match_else**: Suggests using `if let` instead of `match` with a single pattern
- **clippy::trait_duplication_in_bounds**: Detects duplicated trait bounds in where clauses
- **clippy::unnecessary_mut_passed**: Warns when unnecessary mutable references are passed to functions

Rustc Lints

- **nonstandard_style**: Warns about non-standard naming and style conventions
- **rust_2018_idioms**: Enables lint group for Rust 2018 Edition idioms compliance
- **trivial_numeric_casts**: Warns about numeric casts that don't change the value
- **unused_lifetimes**: Warns about unused lifetime parameters
- **unused_qualifications**: Warns about unnecessary type qualifications

Warning Level

- **warnings**: Treats all warnings as errors to prevent problematic code from being accepted

</details>

+++ crates/lib/core/tests/crypto/sha256.rs
@@ -13,10 +13,10 @@ fn sha256_hash_bytes() {
     let ifelts = [
         group_slice_elements::<u8, 4>(&[ibytes.clone(), ipadding].concat())
             .iter()
-            .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+            .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
             .rev()
             .collect::<Vec<u64>>(),
-        vec![length_in_bytes as u64; 1],
+        vec![length_in_bytes; 1],
     ]
     .concat();
 
@@ -62,7 +62,7 @@ fn sha256_hash_bytes() {
     let obytes = hasher.finalize();
     let ofelts = group_slice_elements::<u8, 4>(&obytes)
         .iter()
-        .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+        .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
         .collect::<Vec<u64>>();
 
     let test = build_test!(source, &[]);
@@ -87,7 +87,7 @@ fn sha256_2_to_1_hash() {
 
     let ifelts = group_slice_elements::<u8, 4>(&ibytes)
         .iter()
-        .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+        .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
         .rev()
         .collect::<Vec<u64>>();
 
@@ -97,7 +97,7 @@ fn sha256_2_to_1_hash() {
     let obytes = hasher.finalize();
     let ofelts = group_slice_elements::<u8, 4>(&obytes)
         .iter()
-        .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+        .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
         .collect::<Vec<u64>>();
 
     let test = build_test!(source, &ifelts);
@@ -116,7 +116,7 @@ fn sha256_1_to_1_hash() {
     let ibytes = rand_array::<Felt, 4>().into_bytes();
     let ifelts = group_slice_elements::<u8, 4>(&ibytes)
         .iter()
-        .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+        .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
         .rev()
         .collect::<Vec<u64>>();
 
@@ -126,7 +126,7 @@ fn sha256_1_to_1_hash() {
     let obytes = hasher.finalize();
     let ofelts = group_slice_elements::<u8, 4>(&obytes)
         .iter()
-        .map(|&bytes| u32::from_be_bytes(bytes) as u64)
+        .map(|&bytes| u64::from(u32::from_be_bytes(bytes)))
         .collect::<Vec<u64>>();
 
     let test = build_test!(source, &ifelts);
